Add json_append and json_remove
This commands cover functionality for lists that are
hard, if not impossible, to do atomically with json_get
and json_set.

Namely, atomically appending and removing elements from a list
within a JSON object is now possible. Note that 'atomically'
here means that the command will fail if the znode's version
changed between get() and set() (it won't retry).

Signed-off-by: Raul Gutierrez S <rgs@itevenworks.net>